% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.TreeMapMatching}
\alias{plot.TreeMapMatching}
\title{Plot a TreeMapMatching object}
\usage{
\method{plot}{TreeMapMatching}(x, y, scale = 1, rgl = FALSE, gg = FALSE, show_matching = TRUE, ...)
}
\arguments{
\item{x}{TreeMapMatching object}

\item{y}{not used}

\item{scale}{upscale tree diameters for a better readability of the plot}

\item{rgl}{bool. plot in 3D with rgl}

\item{gg}{bool. Plot with ggplot instead of base R graphic system}

\item{show_matching}{bool. Disable matching display}

\item{...}{not used}
}
\description{
Plot a TreeMapMatching object
}
