% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{PRF025_Lidar}
\alias{PRF025_Lidar}
\title{Tree Map From Mobile Lidar Data}
\format{
A data frame with 2 rows and 9 variables:
\describe{
  \item{treeID}{Integer. Unique identifier for the measured tree.}
  \item{plotID}{Character. Identifier of the plot where the tree is located.}
  \item{TH}{Numeric. Tree height in meters.}
  \item{DBH}{Numeric. Diameter at breast height in meters.}
  \item{X}{Numeric. X coordinate (e.g., UTM Easting).}
  \item{Y}{Numeric. Y coordinate (e.g., UTM Northing).}
  \item{Z}{Numeric. Elevation above sea level in meters.}
  \item{rmse}{Numeric. Not documented.}
  \item{angle_range}{Numeric. Not documented.}
}
}
\source{
Generated from MLS lidar data.
}
\usage{
PRF025_Lidar
}
\description{
Tree Map From Mobile Lidar Data
}
\examples{
data(PRF025_Lidar)
head(PRF025_Lidar)
}
\keyword{datasets}
