% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize tree data for use in TreeMap matching}
\usage{
standardize(
  data,
  xname,
  yname,
  zname,
  zunits = "cm",
  crs = sf::NA_crs_,
  idname = NULL
)
}
\arguments{
\item{data}{A `data.frame` or `sf` object with tree records .}

\item{xname}{Name of the column containing X coordinates (no used with sf objects).}

\item{yname}{Name of the column containing Y coordinates (no used with sf objects).}

\item{zname}{Name of the column containing the Z dimension. Usually DBH (Diameter at Breast Height)
values or tree height.}

\item{zunits}{Unit of the zname values. Either `"cm"` (default) or `"m"` or `"mm"`. If "cm" or "mm"
the value will be converted to meters.}

\item{crs}{Coordinate Reference System to assign to the output geometry (as a `sf::st_crs()` object).}

\item{idname}{Name of the column containing a unique ID for each tree. Not mandatory but will be
leverage in the plot function to display real tree IDs and to add the id of each tree in the match table}
}
\value{
An `sf` object with:
\itemize{
  \item Geometry from the X and Y columns.
  \item A new column `ZDIM` in meters
  \item CRS set as specified.
  \item An attribute `"standardized" = TRUE`.
}
}
\description{
Converts a table of tree data into a standardized spatial format with consistent naming,
and units and coordinate reference system to work with any user data storage convention.
This function prepares the input data to be used as part of a `TreeMapMatchinh` object for tree
matching and analysis. It also perform some quality check in order to catch obvious error in the
data such as NAs or duplicated trees.
}
\examples{
data(PRF025_Field)
tree_map <- standardize(
  PRF025_Field,
  xname = "Field_Xpj",
  yname = "Field_Ypj",
  zname = "DBH",
  zunits = "cm",
  crs = 2959)
}
\seealso{
\code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_crs}}, \code{\link[dplyr]{rename_with}}
}
