% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.TreeMapMatching}
\alias{plot.TreeMapMatching}
\title{Plot a TreeMapMatching object}
\usage{
\method{plot}{TreeMapMatching}(x, y, scale = 1, rgl = FALSE, ...)
}
\arguments{
\item{x}{TreeMapMatching object}

\item{y}{not used}

\item{scale}{upscale tree diameters for a better readability of the plot}

\item{rgl}{bool plot in 3D with rgl}

\item{...}{not used}
}
\description{
Plot a TreeMapMatching object
}
