% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize tree data for use in TreeMap matching}
\usage{
standardize(data, xname, yname, dbhname, dbhunits = "cm", crs = sf::NA_crs_)
}
\arguments{
\item{data}{A `data.frame` with tree records.}

\item{xname}{Name of the column containing X coordinates.}

\item{yname}{Name of the column containing Y coordinates.}

\item{dbhname}{Name of the column containing DBH (Diameter at Breast Height) values.}

\item{dbhunits}{Unit of the DBH values. Either `"cm"` (default) or `"m"`. If "cm" the value will
be converted to meters.}

\item{crs}{Coordinate Reference System to assign to the output geometry (as a `sf::st_crs()` object).}
}
\value{
An `sf` object with:
\itemize{
  \item Geometry from the X and Y columns.
  \item A renamed column `DBH` in meters
  \item CRS set as specified.
  \item An attribute `"standardized" = TRUE`.
  \item Class modified to include `"TreeMap"`.
}
}
\description{
Converts a table of tree data into a standardized spatial format with consistent naming,
DBH units, and coordinate reference system to work with any user data storage convention.
This function prepares the input data to be used as part of a `TreeMapMatchinh` object for tree
matching and analysis.
}
\details{
The function renames the user-provided DBH column to a standardized `DBH` name and ensures
the unit is in centimeters. If DBH was originally in meters, it is converted by multiplying by 100.
The object returned inherits from class `"TreeMap"`, making it compatible with downstream
functions that expect standardized input.
}
\examples{
data(PRF025_Field)
tree_map <- standardize(
  PRF025_Field,
  xname = "Field_Xpj",
  yname = "Field_Ypj",
  dbhname = "DBH",
  dbhunits = "cm",
  crs = 2959)
}
\seealso{
\code{\link[sf]{st_as_sf}}, \code{\link[sf]{st_crs}}, \code{\link[dplyr]{rename_with}}
}
